migrated harbess tests to scenario
