--- gateway-api-integrator/tests/unit/requirements.txt
@@ -1,4 +1,0 @@
-mock
-pytest
-pytest-cov
-pyyaml
