+++ gateway-api-integrator/tox.toml
@@ -121,7 +121,6 @@ commands = [
     "--tb",
     "native",
     "--ignore={[vars]tst_path}unit",
-    "--ignore={[vars]tst_path}unit_scenario",
     "--log-cli-level=INFO",
     "-s",
     { replace = "posargs", extend = "true" },
